Changed link placeholder to encourage HTTPS
Change link placeholder from http:// to https:// to encourage HTTPS-Links. 

I can imagine that it would be a good idea to even pre-fill the textbox with "https://", instead of just setting the placeholder, so that users don't have to type it out everytime, but this seems like a profound change worthy of discussion in an issue first.